Moved folder lexer from parse to selectively.
