+++ lexer/Token.ts
