+++ lexer/index.spec.ts
