+++ lexer/index.ts
